Check whether the packages is already installed

+++ setup/macos/04-setup-fish.sh
@@ -1,25 +1,29 @@
 #!/usr/bin/env bash
 
-function in_file {
-  PATTERN="$1"
-  FILEPATH="$2"
-  return $(grep -Fxq "$PATTERN" $FILEPATH)
-}
-
-if in_file "$(which fish)" /etc/shells
+# Set fish as default shell
+if grep -Fxq "$(which fish)" /etc/shells
 then
-  echo "$(which fish) is already added to default shells"
+  echo "$(which fish) is already added to default shells."
 else
   sudo sh -c "echo $(which fish) >> /etc/shells"
+  chsh -s "$(which fish)"
 fi
-chsh -s "$(which fish)"
 
-fish -c "curl -sL https://git.io/fisher | source && fisher install jorgebucaran/fisher"
+# Install Fisher
+# https://github.com/jorgebucaran/fisher#installation
+if test -f ~/.config/fish/functions/fisher.fish
+then
+  echo "Fisher is already installed."
+else
+  fish -c "curl -sL https://git.io/fisher | source && fisher install jorgebucaran/fisher"
+fi
 
+# Install Fisher packages
 fish -c "fisher install catppuccin/fish"
 fish -c "fisher install jhillyerd/plugin-git"
 fish -c "fisher install jethrokuan/z"
 fish -c "fisher install franciscolourenco/done"
 fish -c "fisher install jorgebucaran/spark.fish"
 
+# Disable fish greeting message
 fish -c "set -U fish_greeting ''"
